+++ Changes.md
@@ -4,4 +4,4 @@
 
 ## 1.0.0 - 2022-06-17
 
-- Initial release
+- Imported from [Haraka](https://github.com/haraka/Haraka)